+++ lib/onlyoffice_webdriver_wrapper/helpers/chrome_helper.rb
@@ -35,13 +35,11 @@ module OnlyofficeWebdriverWrapper
         credentials_enable_service: false
       }
       caps = Selenium::WebDriver::Remote::Capabilities.chrome
-      caps[:logging_prefs] = { browser: 'ALL' }
       caps[:proxy] = Selenium::WebDriver::Proxy.new(ssl: "#{@proxy.proxy_address}:#{@proxy.proxy_port}") if @proxy
       switches = add_useragent_to_switches(DEFAULT_CHROME_SWITCHES)
       options = Selenium::WebDriver::Chrome::Options.new(args: switches,
                                                          prefs: prefs)
-      webdriver_options = { options: options,
-                            desired_capabilities: caps,
+      webdriver_options = { capabilities: [caps, options],
                             service: chrome_service }
       driver = Selenium::WebDriver.for :chrome, webdriver_options
       maximize_chrome(driver)